agents/triage_engine/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
transformers>=4.35.2
torch>=2.0.0
pydantic>=2.5.0
accelerate>=0.21.0
python-dotenv>=1.0.0
google-generativeai>=0.8.3
websockets>=12.0
groq>=0.4.2